\docType{package}
\name{mlPhaser-package}
\alias{mlPhaser-package}
\title{mlPhaser}
\description{
  \tabular{ll}{ Package: \tab mlPhaser\cr Type: \tab
  Package\cr Version: \tab 0.01\cr Date: \tab 2012-08-28\cr
  Author: \tab Dave T. Gerrard
  <david.gerrard@manchester.ac.uk>\cr License: \tab GPL (>=
  2)\cr LazyLoad: \tab yes\cr }
}
\details{
  mlPhaser

  Some text

  The main functions are: \code{\link{phaseReport}},
  \code{\link{getValidHaploGroups}},

  ...
}

