\docType{methods}
\name{printHaploGroups}
\alias{printHaploGroups}
\title{Print haplotype groups}
\usage{
  printHaploGroups(haploListOfLists)
}
\arguments{
  \item{haploListOfLists}{A list of haplotype groups.}
}
\value{
  Nothing. Just prints.
}
\description{
  Prints haplotype groups from a list.
}
\details{
  Works through a list of haplotype groups (quite a complex
  list structure) and prints out one group per line. For
  presentation only, results cannot be re-used.
}

